+++ demos/rgba.js
@@ -1,0 +1,13 @@
+/**
+ * A helper class for quickly making rgba color strings.
+ *
+ * @date    2026-06-10
+ * @author  Ikaros Kappler
+ * @version 1.0.0
+ */
+
+(function (_context) {
+  _context.rgba = function (r, g, b, a) {
+    return "rgba(" + r + "," + g + "," + b + "," + a + ")";
+  };
+})(globalThis);